pyproject.toml
[project]
name = "OrchestratorX"
version = "0.1.0"
description = "This is an Multi Utility Agentic Chatbot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.10.0",
    "google-genai>=1.63.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "tavily-python>=0.7.21",
]
